+++ src/test/java/com/back/pinco/domain/user/controller/UserControllerTest.java
@@ -1,5 +1,7 @@
 package com.back.pinco.domain.user.controller;
 
+import com.back.pinco.domain.likes.repository.LikesRepository;
+import com.back.pinco.domain.pin.entity.Pin;
 import com.back.pinco.domain.user.entity.User;
 import com.back.pinco.domain.user.repository.UserRepository;
 import com.back.pinco.domain.user.service.UserService;
@@ -17,11 +19,12 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.UUID;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.handler;
 
 @SpringBootTest // 실제 스프링 부트 애플리케이션 컨텍스트 로드
 @AutoConfigureMockMvc(addFilters = false) // MockMvc 자동 구성
@@ -37,6 +40,8 @@ class UserControllerIntegrationTest {
     PasswordEncoder passwordEncoder;
     @Autowired
     UserService userService;
+    @Autowired
+    LikesRepository likesRepository;
 
 
     @Test
@@ -791,6 +796,36 @@ class UserControllerIntegrationTest {
         assertThat(after.getUserName()).isEqualTo(oldName);
         assertThat(passwordEncoder.matches(oldPwd, after.getPassword())).isTrue();
         assertThat(passwordEncoder.matches("NewPassword123!", after.getPassword())).isFalse();
+    }
+
+
+    @Test
+    @DisplayName("특정 사용자가 좋아요 누른 핀 목록 전달")
+    @Transactional
+    void getPinsLikedByUser() throws Exception {
+        // given
+        Long userId = 1L;
+
+        Integer[] pinIds = likesRepository.findPinsByUserIdAndLikedTrue(userId)
+                .stream()
+                .map(Pin::getId)
+                .map(id -> id.intValue())
+                .toArray(Integer[]::new);
+
+        // when & then
+        mvc.perform(
+                        get("/api/user/{userId}/likespins", userId)
+                )
+                .andDo(print())
+                .andExpect(handler().handlerType(UserController.class))
+                .andExpect(handler().methodName("getPinsLikedByUser"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.errorCode").value("200"))
+                .andExpect(jsonPath("$.msg").value("성공적으로 처리되었습니다"))
+
+                .andExpect(jsonPath("$.data").isArray())
+                .andExpect(jsonPath("$.data.length()").value(likesRepository.countByUser_idAndLikedTrue(userId)))
+                .andExpect(jsonPath("$.data[*].id", containsInAnyOrder(pinIds)));;
     }
 
 }
